feat(wasm): Execute block hook sudo dispatches
