+++ x/wasm/module_test.go
@@ -129,7 +129,7 @@ var (
 	addrAcc1     = keyPubAddr()
 	addr1        = addrAcc1.String()
 	testContract = mustLoad("./testdata/hackatom.wasm")
-	maskContract = testdata.ReflectContractWasm()
+	maskContract = testdata.ReflectContractWasm
 	oldContract  = mustLoad("./testdata/escrow_0.7.wasm")
 )
